# Conflicts:
#	packages/cubejs-client-ngx/package.json
#	packages/cubejs-client-ngx/yarn.lock

+++ rust/cubestore/src/remotefs/mod.rs
@@ -4,6 +4,7 @@ pub mod s3;
 
 use crate::config::injection::DIService;
 use crate::di_service;
+use crate::util::lock::acquire_lock;
 use crate::CubeError;
 use async_trait::async_trait;
 use chrono::{DateTime, Utc};
@@ -13,7 +14,7 @@ use log::debug;
 use std::fmt::Debug;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
-use tempfile::NamedTempFile;
+use tempfile::{NamedTempFile, PathPersistError};
 use tokio::fs;
 use tokio::sync::{Mutex, RwLock};
 
@@ -42,6 +43,20 @@ pub trait RemoteFs: DIService + Send + Sync + Debug {
         self.local_file(&format!("uploads/{}", remote_path)).await
     }
 
+    /// Convention is to use this directory for creating files to be uploaded later.
+    async fn uploads_dir(&self) -> Result<String, CubeError> {
+        // Call to `temp_upload_path` ensures we created the uploads dir.
+        let file_in_dir = self
+            .temp_upload_path("never_created_remote_fs_file")
+            .await?;
+        Ok(Path::new(&file_in_dir)
+            .parent()
+            .unwrap()
+            .to_str()
+            .unwrap()
+            .to_owned())
+    }
+
     /// In addition to uploading this file to the remote filesystem, this function moves the file
     /// from `temp_upload_path` to `self.local_path(remote_path)` on the local file system.
     async fn upload_file(&self, temp_upload_path: &str, remote_path: &str)
@@ -62,17 +77,26 @@ pub trait RemoteFs: DIService + Send + Sync + Debug {
 
 #[derive(Debug)]
 pub struct LocalDirRemoteFs {
-    remote_dir_for_debug: PathBuf,
-    remote_dir: RwLock<PathBuf>,
+    remote_dir_for_debug: Option<PathBuf>,
+    remote_dir: RwLock<Option<PathBuf>>,
     dir: PathBuf,
     dir_delete_mut: Mutex<()>,
 }
 
 impl LocalDirRemoteFs {
-    pub fn new(remote_dir: PathBuf, dir: PathBuf) -> Arc<LocalDirRemoteFs> {
+    pub fn new(remote_dir: Option<PathBuf>, dir: PathBuf) -> Arc<LocalDirRemoteFs> {
         Arc::new(LocalDirRemoteFs {
             remote_dir_for_debug: remote_dir.clone(),
             remote_dir: RwLock::new(remote_dir),
+            dir,
+            dir_delete_mut: Mutex::new(()),
+        })
+    }
+
+    pub fn new_noop(dir: PathBuf) -> Arc<LocalDirRemoteFs> {
+        Arc::new(LocalDirRemoteFs {
+            remote_dir_for_debug: None,
+            remote_dir: RwLock::new(None),
             dir,
             dir_delete_mut: Mutex::new(()),
         })
@@ -92,36 +116,86 @@ impl RemoteFs for LocalDirRemoteFs {
         temp_upload_path: &str,
         remote_path: &str,
     ) -> Result<(), CubeError> {
-        debug!("Uploading {}", remote_path);
-        let remote_dir = self.remote_dir.write().await;
-        let dest = remote_dir.as_path().join(remote_path);
-        fs::create_dir_all(dest.parent().unwrap()).await?;
-        fs::copy(&temp_upload_path, dest.clone()).await?;
+        if let Some(remote_dir) = self.remote_dir.write().await.as_ref() {
+            debug!("Uploading {}", remote_path);
+            let dest = remote_dir.as_path().join(remote_path);
+            fs::create_dir_all(dest.parent().unwrap())
+                .await
+                .map_err(|e| {
+                    CubeError::internal(format!(
+                        "Create dir {}: {}",
+                        dest.parent().as_ref().unwrap().to_string_lossy(),
+                        e
+                    ))
+                })?;
+            fs::copy(&temp_upload_path, dest.clone())
+                .await
+                .map_err(|e| {
+                    CubeError::internal(format!(
+                        "Copy {} -> {}: {}",
+                        temp_upload_path,
+                        dest.to_string_lossy(),
+                        e
+                    ))
+                })?;
+        }
         let local_path = self.dir.as_path().join(remote_path);
         if Path::new(temp_upload_path) != local_path {
-            fs::rename(&temp_upload_path, local_path).await?;
+            fs::create_dir_all(local_path.parent().unwrap())
+                .await
+                .map_err(|e| {
+                    CubeError::internal(format!(
+                        "Create dir {}: {}",
+                        local_path.parent().as_ref().unwrap().to_string_lossy(),
+                        e
+                    ))
+                })?;
+            fs::rename(&temp_upload_path, local_path.clone())
+                .await
+                .map_err(|e| {
+                    CubeError::internal(format!(
+                        "Rename {} -> {}: {}",
+                        temp_upload_path,
+                        local_path.to_string_lossy(),
+                        e
+                    ))
+                })?;
         }
         Ok(())
     }
 
     async fn download_file(&self, remote_path: &str) -> Result<String, CubeError> {
-        let local_file = self.dir.as_path().join(remote_path);
+        let mut local_file = self.dir.as_path().join(remote_path);
         let local_dir = local_file.parent().unwrap();
         let downloads_dir = local_dir.join("downloads");
         fs::create_dir_all(&downloads_dir).await?;
         if !local_file.exists() {
             debug!("Downloading {}", remote_path);
-            let remote_dir = self.remote_dir.read().await;
-            let temp_path = NamedTempFile::new_in(&downloads_dir)?.into_temp_path();
-            fs::copy(remote_dir.as_path().join(remote_path), &temp_path)
-                .await
-                .map_err(|e| {
-                    CubeError::internal(format!(
-                        "Error during downloading of {}: {}",
-                        remote_path, e
-                    ))
-                })?;
-            temp_path.persist(&local_file)?;
+            if let Some(remote_dir) = self.remote_dir.write().await.as_ref() {
+                let temp_path =
+                    tokio::task::spawn_blocking(move || NamedTempFile::new_in(downloads_dir))
+                        .await??
+                        .into_temp_path();
+                fs::copy(remote_dir.as_path().join(remote_path), &temp_path)
+                    .await
+                    .map_err(|e| {
+                        CubeError::internal(format!(
+                            "Error during downloading of {}: {}",
+                            remote_path, e
+                        ))
+                    })?;
+                local_file =
+                    tokio::task::spawn_blocking(move || -> Result<PathBuf, PathPersistError> {
+                        temp_path.persist(&local_file)?;
+                        Ok(local_file)
+                    })
+                    .await??;
+            } else {
+                return Err(CubeError::internal(format!(
+                    "File not found: {}",
+                    local_file.as_os_str().to_string_lossy()
+                )));
+            }
         }
         Ok(local_file.into_os_string().into_string().unwrap())
     }
@@ -129,15 +203,16 @@ impl RemoteFs for LocalDirRemoteFs {
     async fn delete_file(&self, remote_path: &str) -> Result<(), CubeError> {
         debug!("Deleting {}", remote_path);
         {
-            let remote_dir = self.remote_dir.write().await;
-            let remote = remote_dir.as_path().join(remote_path);
-            if fs::metadata(remote.clone()).await.is_ok() {
-                fs::remove_file(remote.clone()).await?;
-                Self::remove_empty_paths(remote_dir.clone(), remote.clone()).await?;
+            if let Some(remote_dir) = self.remote_dir.write().await.as_ref() {
+                let remote = remote_dir.as_path().join(remote_path);
+                if fs::metadata(remote.clone()).await.is_ok() {
+                    fs::remove_file(remote.clone()).await?;
+                    Self::remove_empty_paths(remote_dir.clone(), remote.clone()).await?;
+                }
             }
         }
 
-        let _local_guard = self.dir_delete_mut.lock().await;
+        let _local_guard = acquire_lock("delete file", self.dir_delete_mut.lock()).await?;
         let local = self.dir.as_path().join(remote_path);
         if fs::metadata(local.clone()).await.is_ok() {
             fs::remove_file(local.clone()).await?;
@@ -158,11 +233,11 @@ impl RemoteFs for LocalDirRemoteFs {
     }
 
     async fn list_with_metadata(&self, remote_prefix: &str) -> Result<Vec<RemoteFile>, CubeError> {
-        let remote_dir = self.remote_dir.read().await;
+        let remote_dir = self.remote_dir.read().await.as_ref().cloned();
         let result = Self::list_recursive(
-            remote_dir.clone(),
+            remote_dir.clone().unwrap_or(self.dir.clone()),
             remote_prefix.to_string(),
-            remote_dir.clone(),
+            remote_dir.unwrap_or(self.dir.clone()),
         )
         .await?;
         Ok(result)
